+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 - [fix: fix RouterLabelRestTemplateInterceptor add response headers exception with httpclient5.](https://github.com/Tencent/spring-cloud-tencent/pull/1239)
 - [feat: support lossless online and offline](https://github.com/Tencent/spring-cloud-tencent/pull/1254)
 - [feat: support lane router](https://github.com/Tencent/spring-cloud-tencent/pull/1256)
+- [feat: add lane router examples](https://github.com/Tencent/spring-cloud-tencent/pull/1260)
 - [feat:upgrade jackson version.](https://github.com/Tencent/spring-cloud-tencent/pull/1262)